+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add new configuration section: Pi Settings
  - Add to Pi Settings: common ```raspi-config``` options
  - Add to Pi Settings: select pigpiod sample rate ([#386](https://github.com/kizniche/mycodo/issues/386))
+ - Add option to completely disable pigpiod ([#386](https://github.com/kizniche/mycodo/issues/386))
 
 ### Bugfixes